fix(context): fixing currentJid, isGroup, and authorNumber

+++ src/commands/downloader/youtube.ts
@@ -63,6 +63,8 @@ export const YouTubeDownloader: CommandFunc = async (
     collector.start();
     await collector.wait();
 
+    if (!collector.contexts.length && ctx.isGroup && !ctx.getGroup()) return;
+    
     await contextSelect.delete();
     const num = collector.contexts[0].text;
     if (isNaN(num as unknown as number)) {
@@ -135,6 +137,15 @@ export const YouTubeDownloader: CommandFunc = async (
     } else if (mediaFlag && mediaFlag === 'mp4' || mediaFlag === 'video') {
       await ctx.replyWithVideo(buffs, `Downloaded *${info.name}*`);
     }
+  });
+
+  stream.on('error', (err) => {
+    if (err.message === 'leaved') {
+        ctx.client.logger.warn(
+            'YouTube Download for: ' + ctx.currentJid() +
+                'stopped because getting kicked from the group!'
+        );
+    } 
   });
 };
